# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PSYX MAIL GEN API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, suspended, banned]
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
  /admin/users/{id}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Get user details (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
  /admin/users/{id}/action:
    patch:
      operationId: adminUserAction
      tags: [admin]
      summary: Perform admin action on user (suspend, activate, ban, delete, reset-password)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserActionInput"
      responses:
        "200":
          description: Action performed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/announcements:
    post:
      operationId: adminBroadcastAnnouncement
      tags: [admin]
      summary: Broadcast announcement to all users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Announcement sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/reserved-usernames:
    get:
      operationId: adminListReservedUsernames
      tags: [admin]
      summary: List reserved usernames
      responses:
        "200":
          description: Reserved usernames
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReservedUsername"
    post:
      operationId: adminAddReservedUsername
      tags: [admin]
      summary: Add a reserved username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservedUsernameInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservedUsername"
  /admin/audit-logs:
    get:
      operationId: adminGetAuditLogs
      tags: [admin]
      summary: Get audit logs
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"

  # Owner Panel
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Auth schemas
...
    AdminUserListResponse:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserDetail"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AdminUserDetail:
      type: object
      required: [id, username, email, displayName, role, status, storageUsed, emailCount, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
        status:
          type: string
        avatarUrl:
          type: ["string", "null"]
        storageUsed:
          type: integer
        emailCount:
          type: integer
        lastLoginAt:
          type: ["string", "null"]
        createdAt:
          type: string
    AdminUserActionInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [suspend, activate, ban, unban, delete, reset-password]
        newPassword:
          type: ["string", "null"]
...
    AnnouncementInput:
      type: object
      required: [title, message]
      properties:
        title:
          type: string
        message:
          type: string
    ReservedUsername:
      type: object
      required: [id, username, reason, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        reason:
          type: string
        createdAt:
          type: string
    ReservedUsernameInput:
      type: object
      required: [username, reason]
      properties:
        username:
          type: string
        reason:
          type: string
    AuditLogListResponse:
      type: object
      required: [logs, total, page, limit]
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AuditLog:
      type: object
      required: [id, adminId, action, targetType, createdAt]
      properties:
        id:
          type: integer
        adminId:
          type: integer
        adminUsername:
          type: string
        action:
          type: string
        targetType:
          type: string
        targetId:
          type: ["integer", "null"]
        details:
          type: ["string", "null"]
        createdAt:
          type: string

    # Owner schemas
...