# lib/api-zod/src/generated/types/auditLog.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PSYX MAIL GEN API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuditLog {
  id: number;
  adminId: number;
  adminUsername?: string;
  action: string;
  targetType: string;
  /** @nullable */
  targetId?: number | null;
  /** @nullable */
  details?: string | null;
  createdAt: string;
}